chore: move timerSlice
