+++ src/app/redux/features/mini-game/timer/timerSlice.ts
